requirements.txt
amqp==5.1.1
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asttokens==2.0.5
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==22.10.0
autopep8==2.0.0
backcall==0.2.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
bleach==5.0.1
carehare==1.0.2
celery==5.2.7
certifi==2021.10.8
cffi==1.15.1
channels==4.0.0
channels-rabbitmq==4.0.1
channels-redis==4.0.0
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.3
daphne==4.0.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.4
Django==4.1.3
django-celery-results==2.4.0
django-ckeditor==6.4.2
django-cors-headers==3.11.0
django-filter==22.1
django-js-asset==2.0.0
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
dnspython==2.2.1
drf-yasg==1.21.4
entrypoints==0.4
eventlet==0.33.1
executing==0.8.3
fastjsonschema==2.16.2
ffmpeg-python==0.2.0
filelock==3.6.0
flower==1.2.0
future==0.18.2
gevent==22.10.2
greenlet==2.0.1
humanize==4.4.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
ipykernel==6.17.1
ipython==8.2.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
itypes==1.2.0
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.17.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-server==1.23.1
jupyter_client==7.4.5
jupyter_core==5.0.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
kombu==5.2.4
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mistune==2.0.4
msgpack==1.0.4
nbclassic==0.4.8
nbclient==0.7.0
nbconvert==7.2.4
nbformat==5.7.0
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==21.3
pamqp==3.2.1
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.2
prometheus-client==0.15.0
prompt-toolkit==3.0.32
psutil==5.9.4
psycopg2==2.9.3
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
Pygments==2.12.0
PyJWT==2.5.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
python-ffmpeg-video-streaming==0.1.15
python3-openid==3.2.0
pytz==2022.6
pywin32==305
pywinpty==2.0.9
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
redis==4.3.4
requests==2.27.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
Send2Trash==1.8.0
service-identity==21.1.0
shaka-streamer==0.5.1
shaka-streamer-binaries==0.5.2
sib-api-v3-sdk==7.4.0
simplejson==3.17.6
six==1.16.0
sniffio==1.3.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
stack-data==0.2.0
terminado==0.17.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
tornado==6.2
traitlets==5.5.0
Twisted==22.10.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.4.0
tzdata==2022.6
uritemplate==4.1.1
urllib3==1.26.9
vine==5.0.0
virtualenv==20.14.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
whitenoise==6.2.0
widgetsnbextension==4.0.3
wrapt==1.14.1
zope.event==4.5.0
zope.interface==5.5.1
